Add Parser.scala
The parser outputs elements of the abstract syntax tree (AST). I want the type
alias for Label = Option[String], but that can only go inside of an object, so
it's in the ast package object in ast/package.scala. There are several other
case classes in there just because there's not enough code for me to feel like
breaking it up into separate files. Maybe as I write the compiler and need to
add methods and such. For now, it appears to be working.