Fix logic in fib.xand
The loop has to exit if n <= 2, but I had gotten the xand arguments backwards.
Essentially, `xand n two reset_n` was saying

```
if (n -= 2) <= 0
  n += 2
  ...body...
else
  exit
```

Put that way, it's easy to see that the loop was actually executing the body
when n <= 2, not when n > 2. So, had to switch around the parameters to the
initial xand.

Since we are no longer being destructive to n (i.e., `n -= 2` becomes more like
`2 -= n`), had to similarly switch around the resetting logic. `reset_two`
sounds like a weird label name, so I just went with `continue`, heh.