Cleaner `clear` code + shorter pause in Snapshot.scala
Before, I just had a magic constant number of lines to try to "clear the
screen". After much experimentation with other methods, I found out that that
method wasn't so bad after all, at least in URXVT. Made it slightly better by
getting rid of the magic number.

Also noticed that running the simulator was a little too slow, so reduced the
pause time. Still a magic number there, but fuck it.